fixed bugs from refactoring
